Split account and dependents code
