+++ src/http/actions/actions.accounts.ts
@@ -198,43 +198,4 @@ export module AccountActions {
         callback(err);
       });
   };
-  /**
-   * ! Get all dependents for a account by account id
-   * * DanBaDo - 2021/12/19
-   * @param accountID {string}
-   * @param callback {Function}
-   */
-  export const findAllAccountDependents = (
-    accountID: string,
-    callback: Function
-  ) => {
-    const queryString = `
-      SELECT *
-      FROM ${_dependentTableName}
-      INNER JOIN ${_accountDependentRealtionshipTableName} 
-        ON ${_dependentTableName}.id = ${_accountDependentRealtionshipTableName}.dependentID
-      INNER JOIN ${_tableName}
-        ON ${_accountDependentRealtionshipTableName}.accountID = ${_tableName}.id
-      WHERE ${_tableName}.id = ${accountID}
-    `;
-    db.query(queryString, (err, result) => {
-      if (err) {
-        callback(err);
-      }
-      callback(
-        null,
-        result.map(item => {
-          const direction: Direction = JSON.parse(item.direction);
-          return {
-            id: item.id,
-            name: item.name,
-            phone: item.phone,
-            direction: direction,
-            status: item.status,
-            date: item.date,
-          };
-        })
-      );
-    });
-  };
 }
