+++ src/routes/routes.accounts.ts
@@ -62,10 +62,4 @@ export const AccountsRoute = (app: express.Application) => {
    * }
    */
   app.put('/password', AccountsController._updatePassword);
-  /**
-   * * Get dependants for a account
-   * @protected Admin or User with Account ID
-   * @param accountID
-   */
-  app.get('/:accountID/dependents', AccountsController._getDependents);
 };
