+++ src/routes/routes.dependents.ts
@@ -1,0 +1,15 @@
+import express from 'express';
+import {DependentsController} from '../http/controllers/controller.dependents';
+import { PermissionActions } from '../models/types/gen/gen.permissions';
+import { AccessControlMiddelware } from '../http/middlewares/middelware.access.validation';
+
+const _resource = PermissionActions.ACCOUNT;
+
+export const DependentsRoute = (app: express.Application) => {
+  /**
+   * * Get dependants for a account
+   * @protected Admin or User with Account ID
+   * @param accountID
+   */
+  app.get('/:accountID/dependents', DependentsController._getDependents);
+};
